Relocate tests to module dirs
