+++ app/components/VerseOfDay/index.tsx
