+++ app/not-found.test.tsx
@@ -1,10 +1,14 @@
 import { render, screen } from '@testing-library/react';
+import type { ReactNode } from 'react';
 import NotFound from '@/app/not-found';
 import TranslationProvider from '@/app/providers/TranslationProvider';
 
 // Mock next/link for testing environment
 jest.mock('next/link', () => {
-  return ({ children, href }: any) => <a href={href}>{children}</a>;
+  // eslint-disable-next-line react/display-name
+  return ({ children, href }: { children: ReactNode; href: string }) => (
+    <a href={href}>{children}</a>
+  );
 });
 
 describe('NotFound page', () => {
