+++ lib/text/__tests__/languageMappings.test.ts
