+++ lib/text/__tests__/stripHtml.test.ts
@@ -1,0 +1,8 @@
+import { stripHtml } from '@/lib/text/stripHtml';
+
+describe('stripHtml', () => {
+  it('removes tags and collapses whitespace', () => {
+    const result = stripHtml('<p>foo <b>bar</b></p>');
+    expect(result).toBe('foo bar');
+  });
+});
